refactor(shelter): extract useShelterMapSearch hook + unit tests
HomePage's map/search orchestration (controlled camera + refs, location
sync, skip-guard, search firing) moves into a useShelterMapSearch hook so
the trickiest logic is unit-testable instead of only verified manually in
the browser. HomePage drops from ~430 to 129 lines and becomes a thin
presentation component.

Adds hook tests covering: LA-center fallback search, geolocation-then-idle
search, the skip-guard (no re-search when restored bounds match the last
search), re-search when restored bounds differ, and map-area search.
Imports only the light Map/utils + constants paths so the tests need just
a @vis.gl/useMap mock + geolocation stub (no heavy Map/Apollo mocking).
